+++ app/models/analysis.rb
@@ -1,5 +1,19 @@
 class Analysis < ActiveRecord::Base
-  validates :num_data, :numericality => {:only_integer => true, :greater_than => 0}
-  validates :interval, :numericality => {:only_integer => true, :greater_than => 0}
-  validates :state, :inclusion => {:in => %w[ processing completed ]}
+  validate :valid_period?
+  validates :batch_size, :numericality => {:only_integer => true, :greater_than => 0, :message => 'invalid'}
+  validates :state, :inclusion => {:in => %w[ processing completed ], :message => 'invalid'}
+
+  private
+
+  def valid_period?
+    errors.add(:from, 'invalid') unless from
+    errors.add(:to, 'invalid') unless to
+
+    return if errors.messages.include?(:from) or errors.messages.include?(:to)
+
+    unless from < to
+      errors.add(:from, 'invalid')
+      errors.add(:to, 'invalid')
+    end
+  end
 end