+++ db/migrate/20180919111425_remove_num_data_and_interval_from_analyses.rb
@@ -1,0 +1,6 @@
+class RemoveNumDataAndIntervalFromAnalyses < ActiveRecord::Migration
+  def change
+    remove_column :analyses, :num_data
+    remove_column :analyses, :interval
+  end
+end