+++ spec/controllers/analyses/execute_spec.rb
@@ -2,7 +2,7 @@
 require 'rails_helper'
 
 describe AnalysesController, :type => :controller do
-  default_params = {:num_data => 1000, :interval => 100, :state => 'processing'}
+  default_params = {:from => 2.month.ago.strftime('%F'), :to => 1.month.ago.strftime('%F'), :batch_size => 100}
 
   after(:all) { Analysis.destroy_all }
 
@@ -10,7 +10,7 @@ describe AnalysesController, :type => :controller do
     before(:all) do
       RSpec::Mocks.with_temporary_scope do
         allow(AnalysisJob).to receive(:perform_later).and_return(true)
-        @res = client.post('/analyses/learn', default_params)
+        @res = client.post('/analyses', default_params)
         @pbody = JSON.parse(@res.body) rescue nil
       end
     end
@@ -23,18 +23,17 @@ describe AnalysesController, :type => :controller do
   end
 
   describe '異常系' do
-    param_keys = default_params.keys - [:state]
     test_cases = [].tap do |tests|
-      (param_keys.size - 1).times {|i| tests << param_keys.combination(i + 1).to_a }
+      (default_params.keys.size - 1).times {|i| tests << default_params.keys.combination(i + 1).to_a }
     end.flatten(1)
 
     test_cases.each do |error_keys|
       context "#{error_keys.join(',')}がない場合" do
-        selected_keys = param_keys - error_keys
+        selected_keys = default_params.keys - error_keys
         before(:all) do
           RSpec::Mocks.with_temporary_scope do
             allow(AnalysisJob).to receive(:perform_later).and_return(true)
-            @res = client.post('/analyses/learn', default_params.slice(*selected_keys))
+            @res = client.post('/analyses', default_params.slice(*selected_keys))
             @pbody = JSON.parse(@res.body) rescue nil
           end
         end
@@ -50,7 +49,7 @@ describe AnalysesController, :type => :controller do
             allow(AnalysisJob).to receive(:perform_later).and_return(true)
             params = default_params.dup
             error_keys.each {|key| params.merge!(key => 'invalid') }
-            @res = client.post('/analyses/learn', params)
+            @res = client.post('/analyses', params)
             @pbody = JSON.parse(@res.body) rescue nil
           end
         end